Wrote molecule equality operator.  Add KChannel checkpoint test

+++ test/src/CheckpointTest.cpp
@@ -87,7 +87,9 @@ TEST(CheckpointTest, Check_BPTI_TIP3) {
     ulong base_runsteps, Continued_runsteps;
     ulong Continued_true_step;
 #if !GOMC_CUDA
+
     int result = chdir("./test/input/Systems/BPTI_TIP3/Base/");
+    
     if (result){
         std::cout << "System call failed!" << std::endl;
         exit(1);
@@ -105,6 +107,7 @@ TEST(CheckpointTest, Check_BPTI_TIP3) {
         std::cout << "System call failed!" << std::endl;
         exit(1);
     }
+
     Simulation SingleRun("in100.conf");
     SingleRun.RunSimulation();
 
@@ -120,6 +123,8 @@ TEST(CheckpointTest, Check_BPTI_TIP3) {
     BoxDimensions & SingleRun_boxDim = SingleRun.GetBoxDim();
     PRNG & Continued_PRNG = Continued.GetPRNG();
     PRNG & SingleRun_PRNG = SingleRun.GetPRNG();
+    Molecules & Continued_mols = Continued.GetMolecules();
+    Molecules & SingleRun_mols = SingleRun.GetMolecules();
 
     EXPECT_EQ(Continued_ml.operator==(SingleRun_ml), true);
     EXPECT_EQ(Continued_ms.operator==(SingleRun_ms), true);
@@ -127,6 +132,7 @@ TEST(CheckpointTest, Check_BPTI_TIP3) {
     EXPECT_EQ(Continued_velocs.operator==(SingleRun_velocs), true);
     EXPECT_EQ(Continued_boxDim.operator==(SingleRun_boxDim), true);
     EXPECT_EQ(Continued_PRNG.operator==(SingleRun_PRNG), true);
+    EXPECT_EQ(Continued_mols.operator==(SingleRun_mols), true);
 
 
     result = chdir("../../../../..");
@@ -149,6 +155,93 @@ TEST(CheckpointTest, Check_BPTI_TIP3) {
         std::cout << "System call failed!" << std::endl;
         exit(1);
     }
+    
+    result = system("exec rm -r ./SingleRun/SingleRun_*");
+    if (result){
+        std::cout << "System call failed!" << std::endl;
+        exit(1);
+    }
+    result = chdir("../../../..");
+    if (result){
+        std::cout << "System call failed!" << std::endl;
+        exit(1);
+    }
+#endif
+}
+
+TEST(CheckpointTest, Check_K_CHANNEL_TIP3) {
+
+    ulong base_runsteps, Continued_runsteps;
+    ulong Continued_true_step;
+#if !GOMC_CUDA
+
+    int result = chdir("./test/input/Systems/K_CHANNEL_TIP3/Base/");
+    if (result){
+        std::cout << "System call failed!" << std::endl;
+        exit(1);
+    }
+    Simulation base("in.conf");
+    base.RunSimulation();
+    result = chdir("../Continued");
+    if (result){
+        std::cout << "System call failed!" << std::endl;
+        exit(1);
+    }
+    Simulation Continued("in.conf");
+    result = chdir("../SingleRun");
+    if (result){
+        std::cout << "System call failed!" << std::endl;
+        exit(1);
+    }
+
+    Simulation SingleRun("in100.conf");
+    SingleRun.RunSimulation();
+
+    MoleculeLookup & Continued_ml = Continued.GetMolLookup();
+    MoleculeLookup & SingleRun_ml = SingleRun.GetMolLookup();
+    MoveSettings & Continued_ms = Continued.GetMoveSettings();
+    MoveSettings & SingleRun_ms = SingleRun.GetMoveSettings();
+    Coordinates & Continued_coords = Continued.GetCoordinates();
+    Coordinates & SingleRun_coords = SingleRun.GetCoordinates();
+    Velocity & Continued_velocs = Continued.GetVelocities();
+    Velocity & SingleRun_velocs = SingleRun.GetVelocities();
+    BoxDimensions & Continued_boxDim = Continued.GetBoxDim();
+    BoxDimensions & SingleRun_boxDim = SingleRun.GetBoxDim();
+    PRNG & Continued_PRNG = Continued.GetPRNG();
+    PRNG & SingleRun_PRNG = SingleRun.GetPRNG();
+    Molecules & Continued_mols = Continued.GetMolecules();
+    Molecules & SingleRun_mols = SingleRun.GetMolecules();
+
+    EXPECT_EQ(Continued_ml.operator==(SingleRun_ml), true);
+    EXPECT_EQ(Continued_ms.operator==(SingleRun_ms), true);
+    EXPECT_EQ(Continued_coords.operator==(SingleRun_coords), true);
+    EXPECT_EQ(Continued_velocs.operator==(SingleRun_velocs), true);
+    EXPECT_EQ(Continued_boxDim.operator==(SingleRun_boxDim), true);
+    EXPECT_EQ(Continued_PRNG.operator==(SingleRun_PRNG), true);
+    EXPECT_EQ(Continued_mols.operator==(SingleRun_mols), true);
+
+
+    result = chdir("../../../../..");
+    if (result){
+        std::cout << "System call failed!" << std::endl;
+        exit(1);
+    } 
+    result = chdir("./test/input/Systems/K_CHANNEL_TIP3");
+    if (result){
+        std::cout << "System call failed!" << std::endl;
+        exit(1);
+    }
+    result = system("exec rm -r ./Base/Base_*");
+    if (result){
+        std::cout << "System call failed!" << std::endl;
+        exit(1);
+    }
+    result = system("exec rm -r ./Continued/Continued_*");
+    if (result){
+        std::cout << "System call failed!" << std::endl;
+        exit(1);
+    }
+    
     result = system("exec rm -r ./SingleRun/SingleRun_*");
     if (result){
         std::cout << "System call failed!" << std::endl;
